+++ lib/torchvision.rb
@@ -50,6 +50,8 @@ require "torchvision/models/wide_resnet101_2"
 require "torchvision/transforms/compose"
 require "torchvision/transforms/functional"
 require "torchvision/transforms/normalize"
+require "torchvision/transforms/random_horizontal_flip"
+require "torchvision/transforms/random_vertical_flip"
 require "torchvision/transforms/resize"
 require "torchvision/transforms/to_tensor"